aegiscore/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "aegiscore"
version = "0.1.0"
description = "AegisCore: Production-grade defensive cybersecurity AI suite"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "joblib>=1.3.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0.0",
    "matplotlib>=3.7.0",
    "shap>=0.43.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.scripts]
aegiscore = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
